# sc2: Adding virtual item tracking for weapon/armour upgrades for performance