… worlds/sc2/pool_filter.py
… worlds/sc2/rules.py
… worlds/sc2/test/test_base.py
… worlds/sc2/test/test_generation.py
… worlds/sc2/test/test_regions.py
… worlds/sc2/test/test_rules.py