## What is this fixing or adding?
Beginning to address the large performance nosedive after PR #5312.

Reported by Mysteryem earlier today in [#ap-core-dev](https://discord.com/channels/731205301247803413/731214280439103580/1413578906912555038).

More can still be done to get the performance back in line, but I'll keep this PR focused on performance boosts from virtual items / overloading `collect()` / `remove()`. (Performance boost from options change now up on PR #5424).

### Weapon/Armour upgrades are now managed by a virtual item
Note there's a slight behavioural change on the logic here: Protoss shield upgrades are no longer covered by logic. As a player I'd say it's fine, shield upgrades are by far the least useful generic upgrades and it was only being checked for in places that were also checking for weapon/armour, so I'd say it's fine and shouldn't lead to any logic holes. The reason shields were left out was because I didn't want to add complexity to hot code to account for the behaviour when the upgrade bundling was set to air/ground -- in that situation, shield upgrades are set to the maximum value of the two upgrade packs.

### Terran defense rating is now regulated by a virtual item
After w/a upgrades were made into virtual functions, the next most expensive rules under the profiler were `terran_competent_comp` and `terran_defense_rating`. `competent_comp` is a far more complex beast, so I aimed for the lower-hanging fruit. I also simplified the logic around it somewhat as it wasn't really matching what I'd expect from a "defense rating". I ran this by the sc2-dev discord chat to make sure it wasn't coming out of left field. The new (and what I expected was old) focus is on specifically items that help you defend even if you F2 your army. This means buildings and units that deploy/siege somehow.

This meant some AA tools and splash tools were removed, like Banshees, splash vikings and valkyries. I looked through all the usages of the function to make sure they still made sense under the new rules and had a reasonable number of permissible combinations of units even under vanilla_items_only.

This refactor and simplification only saved about 2s out of ~220s in my 5-yaml setup under cProfile. There's a few other low-hanging fruit, like terran power level, but I think it's safe to say the payoff is probably too small to promise it as part of this PR.

## How was this tested?
* Ran all unit tests.
* Ran a test generation with 5 template yamls. Generation time went from ~89s to ~68 seconds
* Ran many generations with 5 template yamls under cProfile. Checked time savings were really coming from the functions that were refactored

## If this makes graphical changes, please attach screenshots.
None