@Mysteryem 2025-09-07
I think now would be a good time to replace the reinventing of the wheel that is `ValidInventory`, and to just use `CollectionState`. That way, any `collect/remove` overrides on the world are supported effectively for free without having to reimplement them for `ValidInventory`, and any future `LogicMixin` would also be supported effectively for free.

I have implemented this on current main in https://github.com/Mysteryem/Archipelago-ahit/commit/2b8c845699bcf789ef92734cf3e55dd7f687d480

One subtle difference is that `CollectionState.__init__()` collects all items in `multiworld.precollected_items`. If that needs to not happen, then `ValidInventory.__init__()` could empty `self.logical_inventory`/`self.prog_items[world.player]` after the `super().__init__()` call.