@MatthewMarinets 2025-09-07
> The thing is that having a usable ranged splash unit (even if mobile) makes defensive missions easier. In missions like The Dig and All-in you barely ever F2

If that's a concern, it should be on the logic for those missions in particular. Mobile units aren't defense score, it's not what players will expect it to mean when they ask about logic, and it's not what devs will expect when they reach for a pre-made logic function to fill in logic for a new location. The usage I'd expect as a dev/logic writer is "Bunkers are really helpful in this mission, but I don't want to require one particular item; let me use this grab-bag of bunker-like things".

All-In logic already checks things like Valkyries in Air version and Banshee/Battlecruiser on All-In ground. The Dig checks for competent AA and common units. The mobile part is already accounted for, it should not be factored into domain-specific functions. I think this is already handled.