@MatthewMarinets 2025-09-07
> I think now would be a good time to replace the reinventing of the wheel that is `ValidInventory`, and to just use `CollectionState`. That way, any `collect/remove` overrides on the world are supported effectively for free without having to reimplement them for `ValidInventory`, and any future `LogicMixin` would also be supported effectively for free.
> 
> I have implemented this on current main in [Mysteryem@2b8c845](https://github.com/Mysteryem/Archipelago-ahit/commit/2b8c845699bcf789ef92734cf3e55dd7f687d480)
> 
> One subtle difference is that `CollectionState.__init__()` collects all items in `multiworld.precollected_items`. If that needs to not happen, then `ValidInventory.__init__()` could empty `self.logical_inventory`/`self.prog_items[world.player]` after the `super().__init__()` call.

I agree that the current `ValidInventory` is a bit of a hack (that class has been around since time immemorial and I assume it's just that way because it predates core's current class setup). But I'm uncomfortable refactoring that in this PR.

I'm not very familiar with working with core's `CollectionState` directly, and whenever I have looked into it I've seen indication of undesirable side-effects like the pre-collected dynamic you mention and whatever `prevent_sweep` is doing. Inheriting directly also means we're susceptible to new/changed side-effects if core ever changes. The current setup is a functionality whitelist rather than a blacklist, and I think that's appropriate for this code. As for logic mixins, I think we used to have one for sc2, no one understood it, Berserker commented it was useless and fewer games should use it, and so we removed it. To me, that indicates the trend is towards supporting _fewer_ `LogicMixin`s, not more.

As a programmer, I'm also just very against inheritance in general. Maybe it's the hellish experience of seeing what Java programmers can do to a C codebase to make it more "object oriented". If there's an option that avoids inheritance, I'm taking it (and Python >= 3.10 typing provides this handy-dandy `Protocol` class that lets me get all the perks of interfaces with none of the downsides).

This isn't the only place we use a protocol-like trick with collection state; there's `TestInventory` in test_rules.py. If a refactor of `ValidInventory` finds some good improvements/new patterns that we want, it should be ported there as well. But that's a case of us wanting the whitelist pattern as well; if someone uses a previously-unused CollectionState function in rules.py, we want to manually track that in `TestInventory` to make sure we're adequately checking if everything is progression properly. And that test itself probably wants a refactor (it's currently more of a fuzzer and is quite slow; I've been advocating for more of an AST-parse approach to hopefully speed it up).

Given this PR has already uncovered an uncomfortable number of bugs, I'd rather keep it focused on fixing those and improving performance without throwing in a refactor that may introduce new bugs. I'm saving a link to your commit so we can use it as a jumping-off point when that refactor happens (next dev cycle, Ziktofel has already slated it as a maintenance-focused one).

tl;dr I agree the class needs a refactor, but won't put it in this PR. Next sc2 dev cycle is maintenance-focused, will do the refactor there.