@Snarkie 2025-09-09
Tested this locally, the generation time was reduced by about 26% avg on my end. Refer to 
https://docs.google.com/spreadsheets/d/1jijO6xtJgNhAriOd0obWX6b-388gdSzuRHbl8nvFWJM
for more details.