@Ziktofel 2025-09-10
Re shields: They can be safely ignored if the option is set to bundle_unit_class as the armor check will happen anyways and shield level can't be any worse under these conditions (attach a comment explaining that). In any other mode the pump up/down for virtual items will work